[TH] implement schema migrations
With this, we add a transaction attribute (thanks 
Chris!) when we migrate. We can then compare 
potential schema files against what we have already 
migrated and apply missing ones.

Because we have to install an attribute on the 
schema for the migration by default, we now also 
handle the partition creation that normally has 
appeared at the top of our initial-schemas.

Our initial-schema files should be renamed to 
include the date time stamp at the front of the 
filename so they get migrated just like any other 
migration file.

We’ll need to export and re-import existing databases 
under this scheme.

This by itself will not kick off any migrations 
unless the initialize database is true, in which
case it will run them all after creating the database.
Alternatively, calling  `(run-migrations) will run 
any that haven’t been applied.