[EN] The fix of a lifetime.
Ok, I found a difference between Datomic in memory and the production
version. In the in-memory version, getting a many value from an entity
gives you a PersistentSet. But in production, it gives you a
java.util.HashSet. So there you go. Can't use `coll?` in a datomic
database function.