[WM] Put connection in atom
It's supposed to be cached so that calling datomic.api/connect
repeatedly is OK. But we were seeing ~5 second pauses sometimes and
traced it down to datomic.api/connect. Putting it into an atom seems to
fix that.

We'll keep working with Cognitect support to figure out what's up, but I
guess there's no reason not to keep doing it this way.